Create lab2_20
Write a complete C++ program which would offer a menu of choices to user regarding what type of food they would like to order. Having taken the choice of the user on food, the program would ask how many of that order the user wants and calculate the total price that needs to be paid. The food menu and the sample run of the program is given below: