Create lab3_13
Write a complete C++ program that gets a number greater than 40 from user. Having done so the
program would calculate the sum of all even numbers and the sum of all odd numbers starting from 0 up
to the entered number. The sum of even and odd numbers would be displayed as the output.