Update Lab6_4
Write a complete C++ program that gets the 5 numbers from user. Then, ask the user to enter a key number in order to search and find it. If the given key is found, display to the user as “Key is found”. Otherwise, display to the user as “Key is not found”.