Create lab2_4
Assume that you are asked to write a program for math students in order to demonstrate how to
calculate volume and area of specific 3D geometric shapes. Write a complete C++ program that offers a
menu to the user and get relevant data according to the selected shape. Having done so, the program
would calculate either the volume or area of the selected shape and display this as the output. Hint: you
can solve this question using if/else statement or with switch…case.