Create lab2_14
Assume that you want to create an algorithm for a toy seller company. The company sells only one
type of toy which is “Spider-Man toy robot” and the price is 20$. Write a complete C++ program to get
the number of Spider-Man toy robot” that user want to buy and display total price that user have to pay.
If total price is more than 100 make ten percent discount on total price.