Create lab3_5
Write a complete C++ program that shows the following menu given in the sample run. Once the user selects a specific type of room, the program would prompt the user how many days they want to reserve the selected room. Once this data is taken from the user, the program would calculate the total amount needed to be paid by the user. Notice that the room prices are in terms of dollar. Hint: you can solve this question using if/else statement or with switch…case.