Create more_questions
Write a complete C/C++ program that will ask the users to select a choice from the list of shapes for area calculation. The menu will appear continuously for letting the user to select a choice until the user enter 'x' from the keyboard for the termination. Having selected a valid choice, the users will be asked for required parameters for the selected shape such as weight, height, etc. Once these values are entered, the program would display the calculated area of the selected shape (Recall that: area of a rectangle = length x width, area of a square= 4 x length, and area of circle is 3.14 x radius"). If an invalid choice is selected on the menu, then the program will display "Sorry, your choice does not exist on the list message on the screen. A sample output of the program is given below: (30 Points) R- Rectangle (in meters) S- Square (in meters) C- Circle (in meters) Press x to exit Which operation would you like to select? R What is length? 80 What is width? 40 Area of a rectangle is 3200 R- Rectangle (in meters) S- Square (in meters) C- Circle (in meters) Press x to exit Which operation would you like to select?C